requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.32.3
requests==2.32.3